+++ package.js
@@ -1,8 +1,8 @@
 Package.describe({
-  name: 'faisalman:ua-parser-js',
+  name: 'evenius:oldname-ua-parser-js',
   version: '2.0.0-beta.2',
-  summary: 'Lightweight JavaScript-based user-agent string parser',
-  git: 'https://github.com/faisalman/ua-parser-js.git',
+  summary: 'Lightweight JavaScript-based user-agent string parser without breaking name changes',
+  git: 'https://github.com/evenius/oldname-ua-parser-js.git',
   documentation: 'readme.md'
 });